pyproject.toml
[project]
name = "embedding-suite"
version = "0.1.4"
description = "A simple, unified interface for generating embeddings from various providers."
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.60.1",
    "python-dotenv>=1.0.1",
]
